Refactored
Refactored the codes
