+++ src/components/NavBar/NavbarElements.js
@@ -22,7 +22,7 @@ export const NavLink = styled(Link)`
   height: 100%;
   cursor: pointer;
   &.active {
-    color: #15cdfc;
+    color: #4f359bff;
   }
 `;
